refactor(dsh-plugin): 拆出通用平台 UI Kit（provider 抽象）+ capi RPC 通道
Booker 方向：对齐 Supabase Platform Kit —— 右侧面板拆成可复用组件库，
底层通过可配置 provider 只关心输入输出；默认用 MCP capi 实现云 API 调用。

Phase 1（本次）：
- 新增 src/client/kit/ 通用层（无云依赖）：
  - provider.ts：PlatformProvider 输入输出协议（= CloudBaseData 泛化）+ KIT_EVENTS 跨组件事件
  - components/ResourceTable.tsx：通用数据表格（排序/分页/复制/CSV）
  - components/UrlPreview.tsx：通用 Webview（URL 输入 + iframe + 最近访问 chips）
  - components/EnvSelect.tsx：通用环境选择器（provider + onSwitched 回调）
  - lib/recent-deploys.ts：最近访问地址（localStorage + KIT_EVENTS.recentDeploys）
  - index.ts：kit 统一导出
- 兼容层（现有引用不变）：
  - PreviewTab → re-export kit UrlPreview
  - EnvSelector → EnvSelect 云版包装（注入 appendToSession 反向联动）
  - DataTableCard → ResourceTable 薄包装
- capi RPC 通道（provider 可切到 MCP capi 直调云 API）：
  - data-service.capi(service, action, params) → bridge.callCloudApi
  - remote-service @Remote('capi') + typert manifest
  - client typert.capi + CloudBaseData.capi 协议

真机验证：kit 重构后 5 tab / 预览 webview / 环境选择器均正常，布局 280/500/500。
22/22 测试 + typecheck + build 全过。

后续（Phase 2）：kit 可独立成包 @cloudbase/ui-kit + 更多通用组件（StorageList/
AuthPanel/MetricsGrid）+ provider 默认实现切到 capi。
