+++ dsh-plugin/src/client/kit/provider.ts
@@ -1,0 +1,26 @@
+import type { CloudBaseData } from "../../shared/types.js";
+
+/**
+ * 通用平台数据 Provider —— kit 的输入输出协议。
+ *
+ * 只关心输入输出，不关心底层实现：调用方可以是腾讯云 API / MCP capi / 其他云
+ * 的 SDK / 纯前端 mock。任何满足该接口的对象都能驱动 kit 组件渲染。
+ *
+ * 当前默认实现 = cloudbase-data 服务（host 端包装 cloudbase-mcp，经 typert RPC
+ * 暴露给浏览器）；kit 组件一律通过 PlatformProvider 访问数据，不做云耦合。
+ */
+export type PlatformProvider = CloudBaseData;
+
+/** kit 跨组件事件名（纯 UI 层信号，与具体云无关）。 */
+export const KIT_EVENTS = {
+  /** 当前环境已切换（会话 set_env 或面板下拉选择）。detail = envId */
+  envBound: "cloudbase-dsh:env-bound",
+  /** 出现新访问地址，应激活预览视图。detail = url */
+  activatePreview: "cloudbase-dsh:activate-preview",
+  /** 最近访问地址列表变化。detail = RecentDeploy[] */
+  recentDeploys: "cloudbase-dsh:recent-deploys",
+} as const;
+
+export type KitEventName = (typeof KIT_EVENTS)[keyof typeof KIT_EVENTS];
+
+export { type CloudBaseData };
